backend added

+++ start.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 echo "welcome server..."
 npm install
-if [ $(docker container ls -q --filter name=my_app_backend_container) != '']; then
+if [ $(docker container ls -a -q --filter name=my_app_backend_container) != '']; then
  docker container stop my_app_backend_container
  docker container rm my_app_backend_container 
 fi
 
-if [ $(docker image ls -q --filter reference=my_app_backend) != '' ];then
+if [ $(docker image ls -a -q --filter reference=my_app_backend) != '' ];then
   docker image rm my_app_backend
 fi
